FIXED: temporary GetAllDependencies fix
ADDED: formula display in GetAllDependencies

NOTE: added an example where formula goes from
NOTE: spreadsheet A to B to A. The dependency
NOTE: stack was erased when fetching from low
NOTE: lowest level A. The fix keeps a copy of
NOTE: the dependency stack around. a proper fix
NOTE: will not use a stack stored in the SS
NOTE: object.
NOTE: the two previous examples of GetAllDependencies
NOTE: worked because case 1 was in the same SS
NOTE: and case 2 had a cyclic exception